add img, fixed day
